+++ src/Upload.ts
@@ -5,7 +5,9 @@ export interface FileUpload {
   fieldName: string;
   mimetype: string;
   encoding: string;
+
   createReadStream(options?: ReadStreamOptions): ReadStream;
+
   capacitor: WriteStream;
 }
 
@@ -16,35 +18,13 @@ export class Upload {
   file?: FileUpload;
 
   constructor() {
-    /**
-     * Promise that resolves file upload details. This should only be utilized
-     * by {@linkcode GraphQLUpload}.
-     * @type {Promise<import("./processRequest.mjs").FileUpload>}
-     */
     this.promise = new Promise((resolve, reject) => {
-      /**
-       * Resolves the upload promise with the file upload details. This should
-       * only be utilized by {@linkcode processRequest}.
-       * @param {import("./processRequest.mjs").FileUpload} file File upload
-       *   details.
-       */
       this.resolve = (file) => {
-        /**
-         * The file upload details, available when the
-         * {@linkcode Upload.promise} resolves. This should only be utilized by
-         * {@linkcode processRequest}.
-         * @type {import("./processRequest.mjs").FileUpload | undefined}
-         */
         this.file = file;
 
         resolve(file);
       };
 
-      /**
-       * Rejects the upload promise with an error. This should only be
-       * utilized by {@linkcode processRequest}.
-       * @param {Error} error Error instance.
-       */
       this.reject = reject;
     });